# Add more BTreeMap like apis to LiteMap