--- a/utils/litemap/benches/litemap.rs
+++ b/utils/litemap/benches/litemap.rs
@@ -59,11 +59,12 @@ fn overview_bench(c: &mut Criterion) {
     bench_lookup(c);
     bench_lookup_large(c);
     bench_from_iter(c);
-    bench_from_iter_large(c);
+    bench_from_iter_rand_large(c);
     bench_from_iter_sorted(c);
     bench_from_iter_large_sorted(c);
-    bench_extend(c);
-    bench_extend_via_litemap(c);
+    bench_extend_rand(c);
+    bench_extend_rand_dups(c);
+    bench_extend_from_litemap_rand(c);
 }
 
 fn build_litemap(large: bool) -> LiteMap<String, String> {
@@ -100,7 +101,7 @@ fn bench_deserialize_large(c: &mut Criterion) {
 }
 
 fn bench_from_iter(c: &mut Criterion) {
-    c.bench_function("litemap/from_iter/small", |b| {
+    c.bench_function("litemap/from_iter_rand/small", |b| {
         let mut ff = build_litemap(false).into_iter().collect::<Vec<_>>();
         ff[..].shuffle(&mut rand::thread_rng());
         b.iter(|| {
@@ -110,8 +111,8 @@ fn bench_from_iter(c: &mut Criterion) {
     });
 }
 
-fn bench_from_iter_large(c: &mut Criterion) {
-    c.bench_function("litemap/from_iter/large", |b| {
+fn bench_from_iter_rand_large(c: &mut Criterion) {
+    c.bench_function("litemap/from_iter_rand/large", |b| {
         let mut ff = build_litemap(true).into_iter().collect::<Vec<_>>();
         ff[..].shuffle(&mut rand::thread_rng());
         b.iter(|| {
@@ -141,8 +142,8 @@ fn bench_from_iter_large_sorted(c: &mut Criterion) {
     });
 }
 
-fn bench_extend(c: &mut Criterion) {
-    c.bench_function("litemap/extend/large", |b| {
+fn bench_extend_rand(c: &mut Criterion) {
+    c.bench_function("litemap/extend_rand/large", |b| {
         let mut ff = build_litemap(true).into_iter().collect::<Vec<_>>();
         ff[..].shuffle(&mut rand::thread_rng());
         b.iter(|| {
@@ -157,8 +158,26 @@ fn bench_extend(c: &mut Criterion) {
     });
 }
 
-fn bench_extend_via_litemap(c: &mut Criterion) {
-    c.bench_function("litemap/extend_via_litemap/large", |b| {
+fn bench_extend_rand_dups(c: &mut Criterion) {
+    c.bench_function("litemap/extend_rand_dups/large", |b| {
+        let mut ff = build_litemap(true).into_iter().collect::<Vec<_>>();
+        ff[..].shuffle(&mut rand::thread_rng());
+        b.iter(|| {
+            let mut map: LiteMap<&String, &String> = LiteMap::with_capacity(0);
+            for _ in 0..2 {
+                let mut iter = ff.iter().map(|(k, v)| (k, v));
+                let step = ff.len().div_ceil(10);
+                for _ in 0..10 {
+                    map.extend(iter.by_ref().take(step));
+                }
+            }
+            black_box(map)
+        })
+    });
+}
+
+fn bench_extend_from_litemap_rand(c: &mut Criterion) {
+    c.bench_function("litemap/extend_from_litemap_rand/large", |b| {
         let mut ff = build_litemap(true).into_iter().collect::<Vec<_>>();
         ff[..].shuffle(&mut rand::thread_rng());
         b.iter(|| {