Add BTreeMap-like APIs to LiteMap to get closer to the std lib BTreeMap, which has familiarity benefits and allows easier integration of LiteMap into existing projects. This is a follow up of https://github.com/unicode-org/icu4x/pull/5894 

* Add `lm_extend` to the traits. The vec implementation attempts have O(N) performance on the best case and an asymptotic worst case of O(NLog(N)), effectively avoiding quadratic worst cases. The end goal is to use this in the deserialization code path, which will also come in a follow-up PR.
* Implements `Extend` for `LiteMap`. This is based on `lm_extend`.
* Prevent quadratic worst case in the existing FromIter implementation by using the newly added `lm_extend`
* Even though `try_get_or_insert` exists, the Entry API is a more familiar API with good performance, as it may avoid a potentially wasted second binary search.