@sffc 2025-02-05
I'm all in favor of `Entry` APIs

I've pushed back for a long time on a generic `extend` function because it is a performance footgun; I think it is even worse than `O(N^2)` in the worst case; could be `O(N^2 log N)` (because there are `N` insertions, and each insertion is `N log N`, with `log N` for the lookup and `N` for the array insert).

`FromIterator` is `O(N log N)` which is the optimal complexity: it collects things into a Vec and then sorts the Vec, and it uses the same Vec for its eventual storage.

We have `extend_from_litemap` which does the more efficient `O(N)` thing since it knows the input is sorted, and I would support adding `extend_from_btreemap` since we can make a similar assumption there. But I lean toward not having a `O(N^2 log N)` generic `extend` function.