@Manishearth 2025-02-05
I agree around `extend`, it could be `extend_sorted()` which takes an ExactSizeIterator, perhaps, but that's about it.