@sffc 2025-02-05
I see, I thought FromIterator was `N log N` but it isn't currently implemented that way. Maybe it should be.

https://github.com/unicode-org/icu4x/blob/eaaf49fd296b7a07ecf36fbbd83c982638934648/utils/litemap/src/store/vec_impl.rs#L110