@arthurprs 2025-02-05
I fixed the quadratic behavior and implemented from_iter and extend as described in the first comment. A sorted fast-path with fallback to sort+dedup.


I benchmarked and I find this is a good direction. Let me know if you agree with this extend otherwise I can take it out.

Before
litemap/from_iter/small 460ns
litemap/from_iter/large **1.4s**
litemap/from_iter_sorted/small 49ns
litemap/from_iter_sorted/large 476us

After
litemap/from_iter/small 217ns
litemap/from_iter/large 15ms
litemap/from_iter_sorted/small 46ns
litemap/from_iter_sorted/large 427us
litemap/extend/large 39ms
litemap/extend_via_litemap/large **1.19s**