@arthurprs 2025-02-06
Once this PR is sorted, I'll open another PR to improve the serde deserialization and avoid quadratic worst case. As it stands it's a potential attack vector.