@sffc 2025-02-15
If you re-open with just the Entry API and the benchmarks, I will approve and merge immediately. I just need a bit more time to study your improved extend algorithm for correctness.