@arthurprs 2025-02-15
Yes but somehow I forgot to leave a comment. I'll split the PR in two, one for Entry api and a second for Extend and removing quadratic worst case for FromIter and Deserialization